Move vite-env-vars to basic
