+++ packages/integrations/mdx/test/basic.test.js
@@ -1,5 +1,6 @@
 import { expect } from 'chai';
 import * as cheerio from 'cheerio';
+import { parseHTML } from 'linkedom';
 import { loadFixture } from '../../../astro/test/test-utils.js';
 
 describe('basic - dev', () => {
@@ -48,6 +49,53 @@ describe('basic - build', () => {
 			const $ = cheerio.load(html);
 			expect($('link[href$=".css"]').attr('href')).to.match(/^\/_astro\//);
 			expect($('script[src$=".js"]').attr('src')).to.match(/^\/_astro\//);
+		});
+	});
+
+	describe('vite env vars', () => {
+		it('Avoids transforming `import.meta.env` outside JSX expressions', async () => {
+			const html = await fixture.readFile('/vite-env-vars/index.html');
+			const { document } = parseHTML(html);
+
+			expect(document.querySelector('h1')?.innerHTML).to.contain('import.meta.env.SITE');
+			expect(document.querySelector('code')?.innerHTML).to.contain('import.meta.env.SITE');
+			expect(document.querySelector('pre')?.innerText).to.contain('import.meta.env.SITE');
+		});
+
+		it('Allows referencing `import.meta.env` in frontmatter', async () => {
+			const { title = '' } = JSON.parse(await fixture.readFile('/vite-env-vars/frontmatter.json'));
+			expect(title).to.contain('import.meta.env.SITE');
+		});
+
+		it('Transforms `import.meta.env` in {JSX expressions}', async () => {
+			const html = await fixture.readFile('/vite-env-vars/index.html');
+			const { document } = parseHTML(html);
+
+			expect(document.querySelector('[data-env-site]')?.innerHTML).to.contain(
+				'https://mdx-is-neat.com/blog/cool-post'
+			);
+		});
+
+		it('Transforms `import.meta.env` in variable exports', async () => {
+			const html = await fixture.readFile('/vite-env-vars/index.html');
+			const { document } = parseHTML(html);
+
+			expect(document.querySelector('[data-env-variable-exports]')?.innerHTML).to.contain(
+				'MODE works'
+			);
+		});
+
+		it('Transforms `import.meta.env` in HTML attributes', async () => {
+			const html = await fixture.readFile('/vite-env-vars/index.html');
+			const { document } = parseHTML(html);
+
+			const dataAttrDump = document.querySelector('[data-env-dump]');
+			expect(dataAttrDump).to.not.be.null;
+
+			expect(dataAttrDump.getAttribute('data-env-prod')).to.not.be.null;
+			expect(dataAttrDump.getAttribute('data-env-dev')).to.be.null;
+			expect(dataAttrDump.getAttribute('data-env-base-url')).to.equal('/');
+			expect(dataAttrDump.getAttribute('data-env-mode')).to.equal('production');
 		});
 	});
 });
