+++ packages/integrations/mdx/test/fixtures/basic/astro.config.mjs
@@ -2,5 +2,6 @@ import { defineConfig } from 'astro/config';
 import mdx from "@astrojs/mdx";
 
 export default defineConfig({
+  site: 'https://mdx-is-neat.com/',
   integrations: [mdx()]
 });
