+++ packages/integrations/mdx/test/fixtures/basic/src/pages/vite-env-vars/frontmatter.json.js
@@ -1,4 +1,4 @@
-import { frontmatter } from './vite-env-vars.mdx';
+import { frontmatter } from './index.mdx';
 
 export function get() {
 	return {
