+++ packages/integrations/mdx/test/fixtures/basic/src/pages/vite-env-vars/index.mdx
