--- packages/integrations/mdx/test/fixtures/mdx-vite-env-vars/astro.config.mjs
@@ -1,9 +1,0 @@
-import mdx from '@astrojs/mdx';
-
-export default {
-	site: 'https://mdx-is-neat.com/',
-	markdown: {
-		syntaxHighlight: false,
-	},
-	integrations: [mdx()],
-}
