--- packages/integrations/mdx/test/fixtures/mdx-vite-env-vars/package.json
@@ -1,8 +1,0 @@
-{
-  "name": "@test/mdx-env-variables",
-  "private": true,
-  "dependencies": {
-    "astro": "workspace:*",
-    "@astrojs/mdx": "workspace:*"
-  }
-}
